doc_work/requirements.txt
python-docx==0.8.10
pypdf2==1.26.0
pdf2docx==0.4.6
docx2python==1.27.1

